+++ src/plone/app/blob/adapters/atfile.py
@@ -1,4 +1,4 @@
-from zope.interface import implements
+from zope.interface import implementer
 from zope.component import adapts
 
 from Products.ATContentTypes.interface import IATFile
@@ -6,9 +6,9 @@ from plone.app.blob.interfaces import IBlobbable
 from plone.app.blob.adapters.ofsfile import BlobbableOFSFile
 
 
+@implementer(IBlobbable)
 class BlobbableATFile(BlobbableOFSFile):
     """ adapter for ATFile objects to work with blobs """
-    implements(IBlobbable)
     adapts(IATFile)
 
     def filename(self):