+++ src/plone/app/blob/adapters/file.py
@@ -1,13 +1,13 @@
-from zope.interface import implements
+from zope.interface import implementer
 from zope.component import adapts
 
 from plone.app.blob.interfaces import IBlobbable
 from plone.app.blob.utils import guessMimetype
 
 
+@implementer(IBlobbable)
 class BlobbableFile(object):
     """ adapter for Python file objects to work with blobs """
-    implements(IBlobbable)
     adapts(file)
 
     def __init__(self, context):