+++ CHANGES.rst
@@ -16,6 +16,9 @@ Bug fixes:
 
 - Add a note to monkey.py for when it is possible to remove it.
   [gforcada]
+- Use zope.interface decorator.
+  [gforcada]
+
 
 1.6.2 (2016-05-12)
 ------------------